# More small parse performance improvements