… src/Elm/Parser/Expose.elm
… src/Elm/Parser/Expression.elm
… src/Elm/Parser/File.elm
… src/Elm/Parser/Imports.elm
… src/Elm/Parser/Layout.elm
… src/Elm/Parser/Node.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/State.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/Elm/Parser/Typings.elm
… src/Elm/Parser/Whitespace.elm
… src/List/Extra.elm
… src/Parser/Extra.elm
… tests/Elm/Parser/ImportsTests.elm
… tests/Elm/Parser/TypingsTests.elm
… tests/Elm/Parser/WhitespaceTests.elm