running the benchmarks with eol2 --optimize-speed in chromium

- before:  81955.5 ms to do 100 runs => 819.555 ms/run
- after: 42913.90000000596 ms to do 100 runs => 429.1390000000596 ms/run

(the improvements are pretty consistent now)

Done by

- moving let declarations to module level if it requires no changes
- changing `succeed () |> map (\() -> x)` to `lazy (\() -> succeed x)`
- changing `succeed a |. x` to `map (\() -> a) x`
- small optimizations for e.g. how string literal parts are folded
- ignore reserved words for type names
- change from `Dict` for `operations` to `case` (and actually pre-compute available operations for all precedences)
- shortcutting lower/uppercase character checks
- faster layout primitives

Not all changes are necessarily performance related. Some change ignored unit arguments to `\() ->`. One fixes import alias parsing.