@lue-bird 2024-07-20
We could do some more random things. For example for
```elm
Core.oneOf
    [ Tokens.multiLineStringLiteral
    , Tokens.stringLiteral
    ]
```
we could share the first " at the cost of readability and a tiny performance improvement.

Same in
```elm
Combine.oneOf
        [ functionWithSignature pointer
        , functionWithoutSignature pointer
        ]
```
where technically both share the declared name at the start, so if the first one fails it shouldn't need to be parsed again, or fail both directly if the name is invalid.
Would changes like that be acceptable?

Another thing: `keyword` (which looks ahead for whitespace) is used for tokens that always have layout after them (e.g. port in port module). Should we switch to token / symbol there? Actually so far I've only seen keyword being used with correct layout after.