@jfmengels 2024-07-20
List.filterMap is optimized in a non-merged PR (https://github.com/mdgriffith/elm-optimize-level-2/pull/75). But it does get optimized in elm-review (which is not using eol2, just doing a few similar patches).

Random optimization idea as well (I'll reply to your other questions tomorrow morning): we do a LOT of string concatenations when parsing string literals. But most of the time we can get the contents through a simple get chompedstring. The only times we can't do that is when there's escaping (which is not the common case) and even then we could probably use getChompedString for the rest of the strings, leading to a lot less concatenations.