@lue-bird 2024-07-20
I don't think this will improve anything because literals that don't contain escaped characters are already parsed fast by `chompWhile`.