@jfmengels 2024-07-21
Yeah, we already do that, I only remembered the part where we add one character. Too bad :sweat_smile: 
(And I had to braindump that before going to bed, didn't have time to re-check)


> we could share the first " at the cost of readability but giving us a tiny performance improvement.

I think that's a good idea. I'm honestly confused as to how this even works, because my mental model says that the first `"` character should cause `multimultiLineStringLiteral` to commit and therefore cause to not be evaluated.

For the other location (`functionWithSignature`, ...), I'm less sure. I think it would be good to have, but at least the way we often do things is by passing the argument (through `andThen` or `map` to a new function which redefines. At this point, I believe that having parser defined as constants leads to faster code, therefore redefining the parser inside of a function leads to slower code. The `pointer` argument is something that I'd like to remove already, and I wonder if sharing the parser for the type name is not going to make that even harder.

> Re:keyword

I don't yet understand the differences between symbol/token/keyword. I'm fine with changing to the fastest one. And maybe having some guidance written down somewhere to explain which one to use in general (or in which case).

> I have a commit ready for replacing record type alias constructor calls with fields. Should I add this here or are there other plans (like: "keep them because records might get replaced by variants in v8")

In terms of breaking changes, I think it's fine. Do you know if it's faster? It removes an intermediary function that does the same thing, but maybe that function would get optimized by the JIT. Those are the questions I pondered and didn't know the answer yet.

> do you use any tools to look at performance at the bigger picture?

I don't have any useful tool for this unfortunately, just individual smaller benchmarks like the ones you made (but which I haven't used recently, do you have one somewhere that I could use as a starting point?) and @jiegillet's 100 run benchmark.

> Core.variable is faster than symbol |. chompWhile

I'm confused as to the final results for this, what is finally the fastest? And is this a pattern we use often? I only see it in `many1Spaces` (which is ran very often yes). And :people_hugging: for your ruined day.