@lue-bird 2024-07-21
> Core.variable is faster than symbol |. chompWhile: I'm confused as to the final results for this, what is finally the fastest?

Core.variable is consistently ~15% faster than with chompWhile. I imagine it could even be faster with `oneOf [ variable, succeed "" ]` replacing regular `chompWhile` calls but that I haven't verified.

> do you have one somewhere that I could use as a starting point? and @jiegillet's 100 run benchmark.

I pushed my benchmark code to this branch: https://github.com/lue-bird/elm-syntax/tree/specific-and-full-benchmarks/benchmark

> I don't yet understand the differences between symbol/token/keyword

The doc explains it well: symbol = token and keyword only parses successfully if the character after the symbol is whitespace.

> I'm honestly confused as to how this even works, because my mental model says that the first " character should cause multimultiLineStringLiteral to commit and therefore cause to not be evaluated.

That's the cool thing about `symbol "\"\"\""`. It only commits if the whole thing parses successfully.