@jfmengels 2024-07-22
This has been amazing @lue-bird, I can't believe you've pretty much made the parser twice as fast :exploding_head: 

I don't see any reason not to merge this as it is. Feel free to continue the performance improvements in new PRs, if you still have the performance itch :smile: 

(Note: I've cleaned up the branch and force-pushed it to merge it nicely.)